reorganise LibraryUploadModal files
