+++ jsapp/js/components/library/LibraryUploadModal/index.ts
@@ -1,0 +1,3 @@
+export { default } from './LibraryUploadModal'
+export type { LibraryUploadModalParams } from './LibraryUploadModal'
+export { openLibraryUploadModal } from './openLibraryUploadModal'
